Move createdAt and updatedAt actions to Traits
