+++ src/Comment.php
@@ -10,6 +10,8 @@ namespace Corcel;
 
 class Comment extends Model
 {
+    use CreatedAtTrait, UpdatedAtTrait;
+
     const CREATED_AT = 'comment_date';
     const UPDATED_AT = null;
 
@@ -106,27 +108,5 @@ class Comment extends Model
         }
 
         return $builder;
-    }
-
-    public function setCreatedAt($value)
-    {
-        $field = static::CREATED_AT;
-        $this->{$field} = $value;
-
-        $field .= '_gmt';
-        $this->{$field} = $value;
-
-        return parent::setCreatedAt($value);
-    }
-
-    public function setUpdatedAt($value)
-    {
-        $field = static::UPDATED_AT;
-        $this->{$field} = $value;
-
-        $field .= '_gmt';
-        $this->{$field} = $value;
-
-        return parent::setUpdatedAt($value);
     }
 }
